requirements.txt
tqdm
unstructured
unstructured[pdf]
streamlit
transformers
torch
qdrant-client
langchain[groq]
einops
langchain
langchain-ollama
langchain-text-splitters
langchain-community 
langchain-core
langchain-huggingface
langchain-qdrant
langchain-unstructured
langgraph
pypdf
uvicorn
fastapi
rank_bm25
faiss-cpu
pdf2image
pytesseract